Move host_insights translator into its own package (#2145)
